requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
anyio==4.3.0
apispec==6.6.1
async-timeout==4.0.3
attrs==23.2.0
bech32==1.2.0
bidict==0.23.1
cachetools==5.3.3
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
cosmpy==0.9.2
dataclasses-json==0.6.5
distlib==0.3.8
docx2txt==0.8
ecdsa==0.18.0
exceptiongroup==1.2.1
fastapi==0.96.1
filelock==3.14.0
frozenlist==1.4.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
identify==2.5.36
idna==3.7
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.51
langchain-text-splitters==0.0.1
langsmith==0.1.54
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
mnemonic==0.21
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.4
openai==0.28.0
orjson==3.10.3
packaging==23.2
pipdeptree==2.16.2
platformdirs==3.11.0
pre-commit==3.7.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.20.0
pydantic==1.10.15
Pygments==2.18.0
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-multipart==0.0.5
python-socketio==5.11.2
PyYAML==6.0.1
redis==4.6.0
referencing==0.35.1
reflex==0.4.0
reflex-hosting-cli==0.1.13
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlmodel==0.0.14
starlette==0.27.0
starlette-admin==0.9.0
tabulate==0.9.0
tenacity==8.2.3
tqdm==4.66.4
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
uagents==0.11.1
urllib3==2.2.1
uvicorn==0.20.0
virtualenv==20.26.1
watchdog==2.3.1
watchfiles==0.19.0
websockets==10.4
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
